Chore: move StudySchema
