+++ src/studies/study.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { StudyService } from './study.service';
 import { StudyController } from './study.controller';
 import { MongooseModule } from '@nestjs/mongoose';
-import { Study, StudySchema } from './schemas/study.schema';
+import { Study, StudySchema } from './study.schema';
 
 @Module({
   imports: [
