+++ src/studies/study.schema.ts
