+++ src/studies/study.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { Study, StudyDocument } from './schemas/study.schema';
+import { Study, StudyDocument } from './study.schema';
 
 @Injectable()
 export class StudyService {
